move to typeDef.ts
